+++ main.js
@@ -1,16 +1,16 @@
 // Бургер меню
 import { burgerMenuBtn, burgerMenu } from "./src/vars";
-import { openBurgerMenu, closeBurgerMenu } from "./src/common";
+import { handlBurgerMenu } from "./src/common";
 
-burgerMenuBtn.addEventListener("click", openBurgerMenu);
-burgerMenu.addEventListener("click", closeBurgerMenu);
+burgerMenuBtn.addEventListener("click", handlBurgerMenu);
+burgerMenu.addEventListener("click", handlBurgerMenu);
 
 // обработчик событий на документ/ accordion; faqS
-import { handleDocumentClick } from "./src/common";
+import { handleDocumentClick } from "./src/app/home-page";
 document.addEventListener("click", handleDocumentClick);
 
 // Автоматическая прокрутка карусели
-import { loadProducts } from "./src/common";
+import { loadProducts } from "./src/app/home-page";
 document.addEventListener("DOMContentLoaded", loadProducts);
 
 // //отображение товаров в каталоге
