--- pages/card-item.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="preconnect" href="https://fonts.googleapis.com" />
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-  <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,400;0,700;1,100&display=swap"
-    rel="stylesheet" />
-  <link rel="stylesheet" href="../src/styles/globalStyles/style.scss" />
-  <title>card-item</title>
-</head>
-
-<body>
-  <div class="container">
-    <div class="card-item">
-      <div class="card-item__image">
-        <img src="../src/assets/images/cardImages/dress-1-1.webp" alt="" />
-      </div>
-      <div class="card-item__price">50.00 ₽</div>
-      <div class="card-item__info">
-        <h3 class="card-item__name">Платье женское</h3>
-        <div class="card-item__icons">
-          <img src="../src/assets/images/globalImages/header_shopping-bag-line.svg" alt="" class="card-item__icon" />
-        </div>
-      </div>
-      <div class="card-item__links">
-        <button class="card-item__btn">
-          <a href="#" class="card-item__link">Подробнее<img src="../src/assets/images/globalImages/card-item-arrow.svg"
-              alt="" class="card-item__link-arrow" /></a>
-        </button>
-        <div class="card-item__rating">★</div>
-      </div>
-    </div>
-  </div>
-</body>
-
-</html>
