+++ pages/shirts-catalog.html
@@ -49,7 +49,7 @@
             <input class="header-form__input" type="text" placeholder="Поиск" />
           </form>
           <a class="personal-account" href="#"><img src="../src/assets/images/globalImages/header_avatar-line.svg"
-              alt="" class="header-authorization" /></a>
+              alt="" /></a>
           <a href="./shopping-cart.html"><img src="../src/assets/images/globalImages/header_shopping-bag-line.svg"
               alt="" class="header-basket" /></a>
         </div>
