--- src/styles/elementsPages/button.scss
@@ -1,14 +1,0 @@
-.button {
-    width: 148px;
-    height: 37px;
-    @include fontTextRegular;
-    font-size: $textSizeOne;
-    color: $backgroundColorDark;
-    background-color: $backgroundColorLight;
-    border: 1px solid $backgroundColorDark;
-}
-
-.button:hover {
-    background-color: $backgroundColorDark;
-    color: $backgroundColorLight;
-}
