--- src/styles/elementsPages/container.css
@@ -1,6 +1,0 @@
-.container {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding-right: 15px;
-  padding-left: 15px;
-}/*# sourceMappingURL=container.css.map */
