+++ src/styles/elementsPages/flashSale.scss
@@ -1,18 +1,20 @@
-.buyNow{
+.buyNow {
     max-width: 1200px;
     margin: 0 auto;
     padding-right: 15px;
     padding-left: 15px;
-    padding-top: 5%;
+    padding-top: 10%;
 }
-.buyNow__title{
+
+.buyNow__title {
     color: $textColorTitle;
     @include fontTextRegular;
     font-size: $titleSizeTwo;
     line-height: 87.3%;
     text-align: left;
 }
-.buyNow__span{
+
+.buyNow__span {
     color: $backgroundColorDark;
     @include fontBadScript;
     font-size: $titleSizeTwo;
@@ -23,68 +25,76 @@
 
 .buyNow__carousel {
     overflow: hidden;
-    width: 100%; /* Ширина карусели */
+    width: 100%;
+    /* Ширина карусели */
 }
 
 .buyNow__carousel-inner {
     display: flex;
-    transition: transform 0.5s ease; /* Плавный переход */
+    transition: transform 0.5s ease;
+    /* Плавный переход */
 }
 
 .buyNow__carousel-card {
-    min-width: calc(100% / 3); /* Ширина одной карточки (3 карточки в ряд) */
+    min-width: calc(100% / 3);
+    /* Ширина одной карточки (3 карточки в ряд) */
     padding: 4%;
 }
 
-.buyNow__carousel-currentPrice{
+.buyNow__carousel-currentPrice {
     color: $textColorTitle;
     @include fontTextRegular;
     font-size: $textSizeOne;
     line-height: 16px;
     text-align: left;
 }
-.buyNow__carousel-startPrice{
+
+.buyNow__carousel-startPrice {
     color: $textColor;
     @include fontTextRegular;
     font-size: 12px;
     line-height: 14px;
     text-align: left;
-    margin-left:13px;
+    margin-left: 13px;
     text-decoration: line-through;
 }
+
 .buyNow__carousel-descriptinAndBasket,
-.buyNow__carousel-info{
-    display:flex;
+.buyNow__carousel-info {
+    display: flex;
     align-items: center;
-    justify-content:space-between;
-    padding-top:10px;
+    justify-content: space-between;
+    padding-top: 10px;
 }
-.buyNow__carousel-description{
+
+.buyNow__carousel-description {
     color: rgb(125, 125, 125);
     @include fontTextRegular;
     font-size: 12px;
     line-height: 14px;
     text-align: left;
 }
-.buyNow__carousel-basket{
+
+.buyNow__carousel-basket {
     cursor: pointer;
     border: none;
     text-decoration: none;
     background: transparent;
 }
+
 .buyNow__carousel-svgBasket:hover {
     background-color: $backgroundColorDark;
     border-radius: 10px;
 }
 
-.buyNow__carousel-thoroughBtn{
+.buyNow__carousel-thoroughBtn {
     color: rgb(18, 18, 18);
     @include fontTextRegular;
     font-size: $textSizeOne;
     line-height: 16px;
     text-align: left;
     border-radius: 0px;
-    border:0.5px solid $textColorTitle;
+    border: 0.5px solid $textColorTitle;
     padding: 14px 25px;
     text-align: center;
     text-decoration: none;
@@ -92,121 +102,168 @@
     cursor: pointer;
     transition: background-color 0.3s ease;
 }
+
 .buyNow__carousel-thoroughBtn:hover {
     background-color: $backgroundColorDark;
-    color:white;
-}
-.buyNow__carousel-stars{
-    color:$backgroundColorDark;
+    color: white;
 }
 
-@media screen and (min-width: 1500px) and (max-width: 5000px){
-    .buyNow{
+.buyNow__carousel-stars {
+    color: $backgroundColorDark;
+}
+
+@media screen and (min-width: 1500px) and (max-width: 5000px) {
+    .buyNow {
         max-width: 100%;
         padding: 5%;
     }
-    .buyNow__title, .buyNow__span{
+
+    .buyNow__title,
+    .buyNow__span {
         font-size: 60px;
     }
+
     .buyNow__span {
         margin-left: 400px;
     }
-    .buyNow__carousel-currentPrice{
+
+    .buyNow__carousel-currentPrice {
         font-size: 28px;
     }
-    .buyNow__carousel-startPrice, .buyNow__carousel-description{
+
+    .buyNow__carousel-startPrice,
+    .buyNow__carousel-description {
         font-size: 22px;
     }
+
     .buyNow__carousel-basket img {
-        width: 30px; 
+        width: 30px;
     }
+
     .buyNow__carousel-descriptinAndBasket,
-    .buyNow__carousel-info,.buyNow__carousel-price{
-    padding-top:10px;
+    .buyNow__carousel-info,
+    .buyNow__carousel-price {
+        padding-top: 10px;
+    }
+
+    .buyNow__carousel-thoroughBtn {
+        padding: 16px 25px;
+        font-size: 24px;
+    }
+
+    .buyNow__carousel-stars {
+        font-size: 22px;
+        font-weight: 300;
+    }
 }
-.buyNow__carousel-thoroughBtn{
-    padding: 16px 25px;
-    font-size: 24px;
-}
-.buyNow__carousel-stars{
-    font-size: 22px;
-    font-weight: 300;
-}
-}
-@media screen and (min-width: 700px) and (max-width: 1024px){
-    .buyNow__title, .buyNow__span{
+
+@media screen and (min-width: 700px) and (max-width: 1024px) {
+
+    .buyNow__title,
+    .buyNow__span {
         font-size: 32px;
     }
-    .buyNow__span{
+
+    .buyNow__span {
         margin-left: 210px;
     }
-    .buyNow__carousel-thoroughBtn{
+
+    .buyNow__carousel-thoroughBtn {
         padding: 8px 15px;
         font-size: 12px;
     }
+
     .buyNow__carousel-descriptinAndBasket,
-    .buyNow__carousel-info,.buyNow__carousel-price{
-    padding-top:5px;
+    .buyNow__carousel-info,
+    .buyNow__carousel-price {
+        padding-top: 5px;
     }
+
     .buyNow__carousel-basket img {
-        width: 16px; 
+        width: 16px;
     }
 }
-@media screen and (min-width: 500px) and (max-width: 700px){
-    .buyNow__title, .buyNow__span{
+
+@media screen and (min-width: 500px) and (max-width: 700px) {
+
+    .buyNow__title,
+    .buyNow__span {
         font-size: 25px;
     }
-    .buyNow__span{
+
+    .buyNow__span {
         margin-left: 165px;
     }
-    .buyNow__carousel-startPrice, .buyNow__carousel-description{
+
+    .buyNow__carousel-startPrice,
+    .buyNow__carousel-description {
         font-size: 12px;
     }
-    .buyNow__carousel-thoroughBtn{
+
+    .buyNow__carousel-thoroughBtn {
         padding: 5px 10px;
         font-size: 10px;
     }
+
     .buyNow__carousel-descriptinAndBasket,
-    .buyNow__carousel-info,.buyNow__carousel-price{
-    padding-top:2px;
+    .buyNow__carousel-info,
+    .buyNow__carousel-price {
+        padding-top: 2px;
     }
-    .buyNow__carousel-currentPrice{
+
+    .buyNow__carousel-currentPrice {
         font-size: 12px;
     }
-    .buyNow__carousel-startPrice, .buyNow__carousel-stars, .buyNow__carousel-description{
+
+    .buyNow__carousel-startPrice,
+    .buyNow__carousel-stars,
+    .buyNow__carousel-description {
         font-size: 10px;
     }
+
     .buyNow__carousel-basket img {
-        width: 14px; 
+        width: 14px;
     }
 }
-@media screen and (min-width: 250px) and (max-width: 500px){
-    .buyNow{
+
+@media screen and (min-width: 250px) and (max-width: 500px) {
+    .buyNow {
         padding-top: 10%;
     }
-    .buyNow__title, .buyNow__span{
+
+    .buyNow__title,
+    .buyNow__span {
         font-size: 25px;
         display: flex;
         justify-content: center;
     }
-    .buyNow__span{
+
+    .buyNow__span {
         margin-left: 210px;
     }
-    .buyNow__carousel-startPrice{
+
+    .buyNow__carousel-startPrice {
         font-size: 14px;
     }
-    .buyNow__carousel-description, .buyNow__carousel-stars, .buyNow__carousel-startPrice{
+
+    .buyNow__carousel-description,
+    .buyNow__carousel-stars,
+    .buyNow__carousel-startPrice {
         font-size: 10px;
     }
-    .buyNow__carousel-thoroughBtn{
-        display:none;
+
+    .buyNow__carousel-thoroughBtn {
+        display: none;
     }
-    .buyNow__carousel-currentPrice{
+
+    .buyNow__carousel-currentPrice {
         font-size: 12px;
     }
+
     .buyNow__carousel-basket img {
-        width: 12px; 
+        width: 12px;
     }
+
     .buyNow__carousel-card {
         padding: 5%;
     }
